TASK: Cover bug #5677 with behat tests (tests shortcuts via FrontendNodeController)
