+++ Neos.Neos/Tests/Behavior/Features/FrontendNodeController/ShortcutRedirectOrRendering.feature
@@ -1,0 +1,145 @@
+@flowEntities
+Feature: Test the Fusion rendering for a request on a shortcut node
+  Background:
+    Given using no content dimensions
+    And using the following node types:
+    """yaml
+    'Neos.ContentRepository:Root': {}
+    'Neos.Neos:Content': {}
+    'Neos.Neos:Sites':
+      superTypes:
+        'Neos.ContentRepository:Root': true
+    'Neos.Neos:Document':
+      properties:
+        title:
+          type: string
+        uriPathSegment:
+          type: string
+    'Neos.Neos:Shortcut':
+      superTypes:
+        'Neos.Neos:Document': true
+      properties:
+        targetMode:
+          type: string
+        target:
+          type: string
+    'Neos.Neos:Site':
+      superTypes:
+        'Neos.Neos:Document': true
+    'Neos.Neos:Test.DocumentType':
+      superTypes:
+        'Neos.Neos:Document': true
+    """
+    And using identifier "default", I define a content repository
+    And I am in content repository "default"
+    When the command CreateRootWorkspace is executed with payload:
+      | Key                | Value           |
+      | workspaceName      | "live"          |
+      | newContentStreamId | "cs-identifier" |
+    And I am in workspace "live" and dimension space point {}
+    And the command CreateRootNodeAggregateWithNode is executed with payload:
+      | Key             | Value             |
+      | nodeAggregateId | "root"            |
+      | nodeTypeName    | "Neos.Neos:Sites" |
+
+    When an asset with id "some-asset" and file name "asset.txt" exists with the content "do we need asset shortcut nodes?"
+
+    And the following CreateNodeAggregateWithNode commands are executed:
+      | nodeAggregateId            | parentNodeAggregateId     | nodeTypeName                | initialPropertyValues                                                                                                       | nodeName |
+      | a                          | root                      | Neos.Neos:Site              | {"title": "Node a"}                                                                                                         | a        |
+      | sir-david-nodenborough     | a                         | Neos.Neos:Test.DocumentType | {"uriPathSegment": "david-nodenborough"}                                                                                    |          |
+      | shortcuts                  | sir-david-nodenborough    | Neos.Neos:Test.DocumentType | {"uriPathSegment": "shortcuts"}                                                                                             |          |
+      | shortcut-first-child-node  | shortcuts                 | Neos.Neos:Shortcut          | {"uriPathSegment": "shortcut-first-child", "targetMode": "firstChildNode"}                                                  |          |
+      | first-child-node           | shortcut-first-child-node | Neos.Neos:Test.DocumentType | {"uriPathSegment": "first-child-node"}                                                                                      |          |
+      | second-child-node          | shortcut-first-child-node | Neos.Neos:Test.DocumentType | {"uriPathSegment": "second-child-node"}                                                                                     |          |
+      | shortcut-parent-node       | shortcuts                 | Neos.Neos:Shortcut          | {"uriPathSegment": "shortcut-parent-node", "targetMode": "parentNode"}                                                      |          |
+      | shortcut-selected-node     | shortcuts                 | Neos.Neos:Shortcut          | {"uriPathSegment": "shortcut-selected-node", "targetMode": "selectedTarget", "target": "node://sir-nodeward-nodington-iii"} |          |
+      | shortcut-selected-asset    | shortcuts                 | Neos.Neos:Shortcut          | {"uriPathSegment": "shortcut-selected-asset", "targetMode": "selectedTarget", "target": "asset://some-asset"}               |          |
+      | shortcut-external-url      | shortcuts                 | Neos.Neos:Shortcut          | {"uriPathSegment": "shortcut-external-url", "targetMode": "selectedTarget", "target": "https://neos.io"}                    |          |
+      | sir-david-nodenborough-ii  | a                         | Neos.Neos:Test.DocumentType | {"uriPathSegment": "david-nodenborough-2"}                                                                                  |          |
+      | sir-nodeward-nodington-iii | sir-david-nodenborough-ii | Neos.Neos:Test.DocumentType | {"uriPathSegment": "nodeward-3"}                                                                                            |          |
+
+    And A site exists for node name "a" and domain "http://localhost" and package "Vendor.Site"
+    And the sites configuration is:
+    """yaml
+    Neos:
+      Neos:
+        sites:
+          'a':
+            preset: default
+            uriPathSuffix: ''
+            contentDimensions:
+              resolver:
+                factoryClassName: Neos\Neos\FrontendRouting\DimensionResolution\Resolver\NoopResolverFactory
+    """
+
+    And the Fusion code for package "Vendor.Site" is:
+    """fusion
+    prototype(Neos.Neos:Test.DocumentType) < prototype(Neos.Fusion:Component) {
+        renderer = afx`
+          node: {node.aggregateId}
+        `
+    }
+    """
+
+  Scenario: Match direct node
+    When I dispatch the following request "/david-nodenborough"
+    Then I expect the following response:
+    """
+    HTTP/1.1 200 OK
+    Content-Type: text/html
+    X-Flow-Powered: Flow/dev Neos/dev
+
+    node: sir-david-nodenborough
+    """
+
+  Scenario: Match shortcut nodes
+    When I dispatch the following request "/david-nodenborough/shortcuts/shortcut-first-child"
+    Then I expect the following response:
+    """
+    HTTP/1.1 303 See Other
+    Location: /david-nodenborough/shortcuts/shortcut-first-child/first-child-node
+    X-Flow-Powered: Flow/dev Neos/dev
+
+
+    """
+
+    When I dispatch the following request "/david-nodenborough/shortcuts/shortcut-parent-node"
+    Then I expect the following response:
+    """
+    HTTP/1.1 303 See Other
+    Location: /david-nodenborough/shortcuts
+    X-Flow-Powered: Flow/dev Neos/dev
+
+
+    """
+
+    When I dispatch the following request "/david-nodenborough/shortcuts/shortcut-selected-node"
+    Then I expect the following response:
+    """
+    HTTP/1.1 303 See Other
+    Location: /david-nodenborough-2/nodeward-3
+    X-Flow-Powered: Flow/dev Neos/dev
+
+
+    """
+
+    When I dispatch the following request "/david-nodenborough/shortcuts/shortcut-selected-asset"
+    Then I expect the following response:
+    """
+    HTTP/1.1 303 See Other
+    Location: http://localhost/_Resources/Testing/Persistent/23dae371d1664f1d9cc7dd029b299ea717298103/asset.txt
+    X-Flow-Powered: Flow/dev Neos/dev
+
+
+    """
+
+    When I dispatch the following request "/david-nodenborough/shortcuts/shortcut-external-url"
+    Then I expect the following response:
+    """
+    HTTP/1.1 303 See Other
+    Location: https://neos.io
+    X-Flow-Powered: Flow/dev Neos/dev
+
+
+    """
